+++ globals.py
@@ -1,0 +1,2 @@
+fileName = 'test.csv'
+outputDateFormat = '%Y-%m-%d'
